v1.1.1.1 JS41 added

+++ Acars/FlightData/Flight.cs
@@ -32,8 +32,8 @@ namespace Acars.FlightData
                 new FlightEvent("4B", 5, "Landing lights Off On TakeOff"  , 5 , 5 , (t) => { return (!t.LandingLights && phase == FlightPhases.TAKEOFF); }),
                 new FlightEvent("4C", 5, "Pitch High Below 1500ft on Departure(Radio)"             , 10, 10, (t) => { return ((t.RadioAltitude < 1500) && t.Pitch > 20); }),
                 new FlightEvent("4D", 5, "Gear down above 250 IAS"             , 10, 30, (t) => { return (t.Gear && t.IndicatedAirSpeed > 255); }),
-                new FlightEvent("6A", 1, "Maximum TakeOff Weight Excceded"             , 30, 30, (t) => { return (phase == FlightPhases.TAKEOFF && ((int)t.GrossWeight > LoadedFlightPlan.Aircraft.MTW)); }),
-                new FlightEvent("6B", 1, "Maximum Landing Weight Excceded"             , 30, 30, (t) => { return (phase == FlightPhases.LANDING && ((int)t.GrossWeight > LoadedFlightPlan.Aircraft.MLW)); }),
+                new FlightEvent("6A", 1, "Maximum TakeOff Weight Excceded"             , 30, 30, (t) => { return (phase == FlightPhases.TAKEOFF && (t.GrossWeight > LoadedFlightPlan.Aircraft.MTW)); }),
+                new FlightEvent("6B", 1, "Maximum Landing Weight Excceded"             , 30, 30, (t) => { return (phase == FlightPhases.LANDING && (t.GrossWeight > LoadedFlightPlan.Aircraft.MLW)); }),
                 new FlightEvent("6C", 1, "Maximum Service Ceiling Excceded"             , 30, 30, (t) => { return (!t.OnGround && (t.Altitude > LoadedFlightPlan.Aircraft.Celling)); }),
                 //new FlightEvent("6D", 1, "Maximum Speed Flap Excceded"             , 30, 30, (t) => {
                 //    var flapSetting = LoadedFlightPlan.Aircraft.FlapSettings.OrderBy(x => x.Key).Where(x => x.Key < t.Flaps).FirstOrDefault().Value;
